Reduce ClusterInfo.h
