+++ arangod/RestHandler/RestAdminClusterHandler.cpp
@@ -39,6 +39,7 @@
 #include "Basics/TimeString.h"
 #include "Cluster/AutoRebalance.h"
 #include "Cluster/AgencyCache.h"
+#include "Cluster/AgencyCallback.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/ClusterHelpers.h"
 #include "Cluster/ClusterInfo.h"
