+++ arangod/Utils/CollectionNameResolver.cpp
@@ -24,6 +24,7 @@
 #include "CollectionNameResolver.h"
 
 #include "ApplicationFeatures/ApplicationServer.h"
+#include "Basics/DownCast.h"
 #include "Basics/NumberUtils.h"
 #include "Basics/ReadLocker.h"
 #include "Basics/StringUtils.h"
