+++ tests/IResearch/AgencyMock.cpp
@@ -30,8 +30,10 @@
 #include "Agency/Store.h"
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Basics/ConditionLocker.h"
+#include "Basics/MutexLocker.h"
 #include "Basics/NumberUtils.h"
 #include "Basics/StringBuffer.h"
+#include "Cluster/AgencyCallback.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/ClusterInfo.h"
